Add monthly task reports to downloads page
