+++ pages/downloads/index.js
@@ -1,10 +1,23 @@
 const { page } = require('@asl/service/ui');
+const moment = require('moment');
 const routes = require('./routes');
 
 module.exports = settings => {
   const app = page({
     ...settings,
     root: __dirname
+  });
+
+  app.get('/', (req, res, next) => {
+    req.api('/reports/task-metrics')
+      .then(response => {
+        res.locals.static.reports = response.json.data.map(report => {
+          const end = moment(report.meta.end);
+          return { id: report.id, year: end.format('YYYY'), month: end.format('MMMM') };
+        });
+        next();
+      })
+      .catch(next);
   });
 
   app.get('/', (req, res) => res.sendResponse());
