+++ pages/reporting/routes.js
@@ -1,6 +1,6 @@
 const download = require('./download');
 const details = require('./details');
-const taskMetrics = require('./task-metrics');
+
 module.exports = {
   download: {
     breadcrumb: false,
@@ -11,10 +11,5 @@ module.exports = {
     breadcrumb: false,
     path: '/details',
     router: details
-  },
-  taskMetrics: {
-    breadcrumb: 'taskMetrics',
-    path: '/task-metrics',
-    router: taskMetrics
   }
 };
