--- pages/reporting/task-metrics/content/index.js
@@ -1,18 +1,0 @@
-module.exports = {
-  page: {
-    title: 'Tasks processed by duration and decision',
-    description: `Download a monthly report broken down by task type including:
-
-     * The number of task processing targets exceeded
-     * The number of submitted, approved, returned or outstanding tasks
-     * The mean time taken to process tasks from submission
-     * The mean time taken to process tasks from assignment
-
-    New reports become available on the 1st of each month.
-
-    ## Download Reports`
-  },
-  links: {
-    taskMetrics: 'Tasks processed by duration and decision for {{month}} {{year}} (ZIP)'
-  }
-};
