--- pages/reporting/task-metrics/index.js
@@ -1,26 +1,0 @@
-const moment = require('moment');
-const { page } = require('@asl/service/ui');
-const routes = require('./routes');
-
-module.exports = settings => {
-  const app = page({
-    ...settings,
-    root: __dirname
-  });
-
-  app.get('/', (req, res, next) => {
-    req.api('/reports/task-metrics')
-      .then(response => {
-        res.locals.static.reports = response.json.data.map(report => {
-          const end = moment(report.meta.end);
-          return { id: report.id, year: end.format('YYYY'), month: end.format('MMMM') };
-        });
-        next();
-      })
-      .catch(next);
-  });
-
-  return app;
-};
-
-module.exports.routes = routes;
