--- pages/reporting/task-metrics/routes.js
@@ -1,9 +1,0 @@
-const exporter = require('./routers/export');
-
-module.exports = {
-  export: {
-    path: '/export/:exportId',
-    router: exporter,
-    breadcrumb: false
-  }
-};
